requirements.txt
requests
python-dotenv
openai==1.12.0

#optional local multilingual sentiment
#torch
fastapi
uvicorn
fasttext
